+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Guide: https://keepachangelog.com/en/1.0.0/
 
 ## 2.0.3
 
+- [Core] Change MapboxCommon dependency to use exact versions in SPM and CocoaPods
 - [Core] Align MapboxCommon dependency with 24.3.1
 
 **MapboxCommon**: v24.3.1